refactor: LLM, Embedding interface 제거
